+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.5
+
+- add attributes like,
+- Duration
+- Borderradius
+- Indicator
+- ViewPortFraction for left,right preview
+- click event
+
+
 ## 1.0.4
 
 - Fixes Bugs